Reorder function related to author
